requirements.txt
google-auth==2.3.3
google-api-python-client==2.31.0
requests==2.26.0
sqlite3==3.36.0
email-validator==1.1.3
beautifulsoup4==4.10.0
lxml==4.6.3